moved utils into src folder
